Create TaxCalculate
The Problem Statement is : 

A retail company must file a monthly sales tax report listing the total sales for the month and the amount of state and county sales tax collected. The state sales tax rate is 5 percent and the county sales tax rate is 2.5 percent. Write a program that asks the user to enter the total sales for the month. The application should calculate and display the following:
•	The amount of county sales tax
•	The amount of state sales tax
•	The total sales tax (county plus state)

Solution explained:

First think this in Pseudocode : as we need four variables to be declared and what will be their data type and purpose

so 
1. Declare Real totalsales - Which will store total sales that user inputs
2. Declare Real countyTax - for storing countyTax
3. Declare Real  stateTax - for stateTax store
4. last the totalTax - which will store totaltax

Next which python module we will need for this define that :
1. InputData() - for taking user input
2. calcCounty() - calculates countytax variable
3. calcState() - calculates countystate 
4. calctotal - calculates totaltax
5. printData - to show the Totaltax

Now writing a Pseudo code will look like this

Module main ()
	//Declare local variables 
	Declare Real totalSales
	Declare Real countyTax
Declare Real totalTax 
	Declare Real stateTax
	
//Function calls
	Call inputData(totalSales) //this module takes in the required user input
	Call calcCounty(totalSales, countyTax) //this module calculates county tax
	Call calcState(totalSales, stateTax)
	Call calcTotal(totalTax, stateTax, countyTax)
	Call Display(totalTax, stateTax, countyTax)
End Module

On this base pseudocode, you can now insert respective modules and declaration